New guides + various adjustments to existing content (#1308)
* Correction of the product name for the German version

* New guides

* Tabs configuration for vserver-linux-ssh-server corrected

* Reworked/updated txAdmin guides

Adjustments to content and new screenshots that match the current version of txAdmin

* Fivem Upvotes Guides adjustments

Adjusted the headings and added the section “Upvotes vs Burst Upvotes”, which explains the differences.

* Add conclusion section to guides

* New guide - Webspace: WordPress Plugins

* Minor adjustments - Webspace: Wordpress install

Adjusted headings

* New guide - Webspace: Mails delivery optimization

* Removed ALT:V & RageMP as we no longer offer it in our store

* New guide: Migrate Domain to ZAP-Hosting

* FiveM - Upvotes: Adjustment to the payment information

Made the note clearer again that the use of ZAP Coins is not possible for this product

* Migration of external content to local hosting (Images)

* Migration of external content to local hosting (Youtube integration)

Video frame removed and replaced by a locally hosted image with a link to the YouTube video.

* Youtube component modified

A video is no longer embedded directly, but replaced by an image (thumbnail), which is hosted locally. A click on the entire component redirects to the YouTube video.

* Visual improvement of the image display (YouTube component)

* Added play icon (YouTube component)

* Removed unrelated files

* FiveM: Removed Dunkos vRP section + 301 Redirect for removed content

Declining popularity and lack of use. The existing content is outdated and poorly written, making it no longer relevant or useful.

* Minor adjustment to the forwarding path

* Vein - Firststeps - Dashboard: Replaced incorrect game image

* Welcome Guide - Replace card image source

* Minor adjustments + Sidebar updated

* Resolved incorrect 301 redirects

* update i18n

* tweak: small adjustments to new guides

* fix: wrong screenshot for port forwarding windows

* fix: typo :p

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Jacob <42719082+ThatGuyJacobee@users.noreply.github.com>