New content and various adjustments (#1397)
* Refactor: VPS/Dedicated - phpMyAdmin

Slightly changed and added a note and reference to the required installation of PHP.

* Refactor: Domain - EasyDNS

* Refactor: Domain - ExpertDNS

* Refactor: Domain - ProfiDNS

* Refactor: Gameserver - Tebex

Headings updated to the current standard, minor adjustments made and added game card overview to supported games.

* Tweak: FiveM - Upvotes

Added FiveM Upvotes to the product list. The correct product for the guide is now displayed. Removed the InlineVoucher as it is not applicable for Upvotes.

* Refactor: Discord Bot - Setup

Added Discord Bot as a product. Updated the structure of headings, adjusted and expanded the content and added a conclusion.

* Fix: Minor adjustments to services.map.tsx

* Tweak: Slightly adjusted sidebar structure & guide titles/labels

* New guide: Dedicated - Network issues

* Tweak: Codeblocks styling

Slightly highlight the code blocks with a border so that they stand out more and do not get lost in the content of the text.

* Tweak: Slightly adjusted sidebar structure

* Refactor: Voiceserver - Backups

Headings updated and content adapted to ensure that the current standards are met.

* Refactor: RedM - Own license key

Updated the guide due to the new Cfx.re portal using the same information like in in the FiveM guide.

* New guide: RedM - txAdmin Setup

* New guide: RedM - TxAdmin Interface

* New guide: Rising World - Dashboard

* Create fivem-change-variant.md

* Remove draft

* Archived: CS:GO game section

CS:GO is no longer actively offered and is also no longer being developed. CS2 is the follow-up to CS:GO.

* Archived: SCP Servermod section

The Servermod extension for SCP has been abandoned and will therefore no longer be maintained.

* Tweak: FiveM - Resolve Database Connection

Information regarding the current syntax of the MySQL connection string and otherwise slightly adapted/updated.

* New guide: Gameserver - Scheduled Tasks

* Archived: Guides from FiveM Section

* Refactor: Voiceserver - Query login

* Sidebar updated + Minor tweaks

* update i18n

* tweak: general improvements

* update i18n

* Minor improvements

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Jacob <42719082+ThatGuyJacobee@users.noreply.github.com>