Guides and tweaks (#1408)
* Tweak: Gameserver - Scheduled Tasks

Fixed broken link to Crontab Guru Generator page.

* Refactor: FiveM - Dev Tools

* New guide: RedM - Dev Tools

* New guide: RedM - Upvotes

* Tweak: FiveM - Upvotes

Added the information on the reset cooldown from Cfx.re

* New guide: Day of Defeat: Source - Dashboard

* Minor tweaks to DODS dashboard guide

* New guide: FiveM - Spawnpoints

* New guide: FiveM ESX - Spawnpoints

* New guide: FiveM - Connect to the server

* New guide: RedM - Connect to the server

* New guide: CS 1.6 - Connect to the server

* New guide: CS:S - Connect to the server

* New guide: TF2 - Connect to the server

* New guide: No More Room in Hell - Connect to the server

* New guide: Minecraft - Connect to the server

* New guide: Ark: Surival Evoled - Connect to the server

* Tweak: Gameserver - FTP access

Content added to make it clearer to the reader how to find their FTP login details.

* Tweak: 7D2D - Become admin

Added note that the comment characters must be removed for the record to be valid.

* New guide: CS2 - Connect to the server

* New guide: Battalion 1944 - Connect to the server

* New guide: DODS - Connect to the server

* New guide: Enshrouded - Connect to the server

* Tweak: VPS - Expand disk

Minor adjustments to keep to the current standard in terms of structure.

* Refactor: Dedicated Server - Server reset

Guide restructured to explain two ways of resetting the data. A simple and quick variant for a new installation and the original variant for the complete and final deletion of the data.

* Tweak: Welcome page

Replaced incorrect icons for the domain and webspace cards.

* New guide: Valheim - Connect to the server

* Tweak: FiveM - Prohibited Resources

* New guide: RedM - Prohibited Resources

* Fixed remaining build issues

* update i18n

---------

Co-authored-by: github-actions <github-actions@github.com>