Docusaurus Design Adjustments (#1534)
* Content & TOC

Adjustments made to the width of the content and table of contentcs area to provent fewer line breaks in the TOC which should improve readability.

* Code block

Adjustments made to the code block and limited the maximum height to 500 px so that longer content does not take up too much space on the page.