Adjustments to the metadata description (#1558)
* Adjustment to the metadata description

The metadata is too long and has been shortened. The "- ZAP-Hosting.com documentation" part has been removed because it was cut off due to its length and is superfluous. This allows more characters to be used for the actual content of the description.

* Wrap description values in quotes to prevent syntax errors

Some description fields were missing quotation marks. This could cause YAML parsing or build errors when special characters (e.g. colon, dash, or quotes) appear in the value.  All description values are now wrapped in quotes as a preventive measure to ensure consistent and safe metadata parsing.

* Tweaks